pyproject.toml
[tool.poetry]
name = "travel-recommendation-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
absl-py = "2.1.0"
annotated-types = "0.7.0"
asgiref = "3.8.1"
asyncmy = "0.2.9"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
blinker = "1.8.2"
bs4 = "0.0.2"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
contourpy = "1.2.1"
cycler = "0.12.1"
dnspython = "2.6.1"
flask = "3.0.3"
flask-cors = "4.0.1"
fonttools = "4.53.0"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.7"
immutabledict = "4.2.0"
injector = "0.22.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
kiwisolver = "1.4.5"
markupsafe = "2.1.5"
matplotlib = "3.9.0"
numpy = "2.0.0"
ortools = "9.10.4067"
outcome = "1.3.0.post0"
packaging = "24.1"
pandas = "2.2.2"
pillow = "10.3.0"
protobuf = "5.27.1"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pymongo = "4.8.0"
pymysql = "1.1.1"
pyparsing = "3.1.2"
pysocks = "1.7.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
requests = "2.32.3"
selenium = "4.22.0"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.5"
sqlalchemy = "2.0.32"
tqdm = "4.66.4"
trio = "0.25.1"
trio-websocket = "0.11.1"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
websocket-client = "1.8.0"
werkzeug = "3.0.3"
wsproto = "1.2.0"
langchain-core = "^0.3.12"
weaviate-client = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
